this is the first commit separating logic from presentation as described in the project roadmap. This commit completely separates logic and presentation in the openwebmail.pl file
several subroutines and support files have been added to
allow this transistion. During this transistion both the old
style and the new style will be simultaneously supported. At
the end of the transistion the old files will be removed and
the code base is expected to reduce to 50% of its current size,
as described in the roadmap. Themes will also be possible at
the end of this transistion

HTML::Template is now REQUIRED. It is included in the modules
directory for sites that cannot install it natively. However,
it is recommended that sites install it natively if possible
to take advantage of the speed of the XS version

this new code and all new future code will be licensed under
the BSD License, a copy of which can now be found in
data/openwebmail/doc/LICENSE.txt. This change was made to make
OpenWebMail even more free and available, without the
restrictions of the GPL

this code is more bleeding-edge than typical -current versions
of the past. It is not recommended for production environments.
However, non-production environments are encouraged to install
and test this code. Bug reports are appreciated

the options footer_template_file and header_template_file
have been added to openwebmail.conf

fix bug that user and group "nobody" were assumed to be gid and
uid 65534, which caused files to be created with 65534 for uid.
(thanks to Martin Och, martin.AT.och.cz for his report)




git-svn-id: svn://openwebmail.acatysmoof.com/svn/trunk@294 9f51437f-89f6-0310-a5ff-ff9a3806d2ee