Separate the logic from the presentation in openwebmail-prefs.pl
Add HTML::Template filtering to all templates via ow-shared.pl.
This allows all templates to be custom filtered before the tmpl
values are processed. Right now it is being used to remove
unnecessary whitespace for better looking final html code

All .tmpl pages are now HTML 4.01 scrict.dtd compliant

Stylesheets are no longer processed as templates. This makes
them _true_ stylesheets - finally!

Add preview feature when switching stylesheets in preferences.
Try it, its fun

We need testers! Please install and let us know how it goes for you!

NOTE TO TRANSLATORS:
Heavy modifications to the template files are required during this
transistion to separate logic from presentation. You will find that the
new .tmpl files are not translated. You may translate these files if you
like, however the next step in this transistion is to convert all
translations to the gettext .po format. The conversion to gettext .po
will force the loss of translations again. In short, if you translate
the new .tmpl files, thats great, but you will probably be asked to
translate again when we move to .po.

Please see the roadmap for an in-depth discussion of where the code
is going:

http://openwebmail.acatysmoof.com/dev/roadmap



git-svn-id: svn://openwebmail.acatysmoof.com/svn/trunk@296 9f51437f-89f6-0310-a5ff-ff9a3806d2ee