separate logic from presentation in openwebmail-main.pl
fix bug that some addresses in addressbook would not show
as a friend in the openwebmail-main message listing

all folders can now be sorted by sender or recipient.
Undisclosed recipients are sorted at the end after the
alphanumeric entries

remove single and double quotes surrounding a sender's name in
the message listing, to provide better sorting and display

popup windows now match the selected stylesheet

tweak login and copyright pages to make them closer to
the originals. Login sizes are closer. Copyright page is
now against white, no matter which stylesheet the user
has selected

upgrade HTML::Template to version 2.9, which is now required

add numerous accesskey keyboard shortcuts to the listview

add "Clear" button when the search field is populated. The refresh
button no longer clears search results, it only refreshes

NOTE TO TRANSLATORS:
Heavy modifications to the template files are required during this
transistion to separate logic from presentation. You will find that the
new .tmpl files are not translated. You may translate these files if you
like, however the next step in this transistion is to convert all
translations to the gettext .po format. The conversion to gettext .po
will force the loss of translations again. In short, if you translate
the new .tmpl files, thats great, but you will probably be asked to
translate again when we move to .po.

Please see the roadmap for an in-depth discussion of where the code
is going:

http://openwebmail.acatysmoof.com/dev/roadmap

We need testers! Please install and let us know how it goes for you!



git-svn-id: svn://openwebmail.acatysmoof.com/svn/trunk@302 9f51437f-89f6-0310-a5ff-ff9a3806d2ee