separate logic from presentation in openwebmail-cal.pl
TRANSLATORS: Please see Alex's notes regarding translations
on his previous logic-presentation commits.

Note: This has required minor changes on calpopup.js and
validate.js. Both moved to data/openwebmail/javascript/
and are loaded in the templates with "scr=" (thanks Scott
Mazur for the suggestion and guidance). It was also
necessary to add 3 styles in the css files.

Included time (12:00) on initialization of hereDate in
calpopup.js; the lack of it was making Iceweasel 3.0 get
confused on the start of DST because it assumed 00:00.

Thanks Alex Teslik for the tips and guidance.



git-svn-id: svn://openwebmail.acatysmoof.com/svn/trunk@332 9f51437f-89f6-0310-a5ff-ff9a3806d2ee