separate logic from presentation in openwebmail-send.pl
htmlarea has been replaced with the xinha rich text editor

html composing support is no longer detected based on the
browser user agent string, but rather on a test of the
javascript capability during login. This brings html composing
support to any browser that can support it automatically
(Thanks to Alicia Pérez, aliciapereza.AT.gmail.com for the report)

users may now attach files to messages using URLs. This was
already supported in the code, but there was no user interface
to designate a URL to attach. System admins may disable this
feature by setting the new enable_urlattach configuration setting
to no

the reply and replyall compose modes now preserve inline
message attachments instead of discarding them, enabling
comment on inline images during reply

users are warned if they are trying to delete a message that
is referenced in the body of the composing message

fix bug 43 that messages with read receipts did not cause
a read receipt message to be sent to the sender
(Thanks to Glenn, glenn.AT.mail.txwes.edu for the report)

fix bug 6 that attachments did not contain a unique id associating
them with the message being composed. Multiple emails with
attachments may now be composed simultaneously
(Thanks to Brat, pawel.AT.rckik.krakow.pl for the report)

add login links to session timeout messages (bug 35 feature request)
(Thanks to Bob Puff, bob.AT.nleaudio.com for the suggestion)

update javascript image resizing code during message reading
to wait until image loading is complete before performing resize

separate the DHTML popup logic from the calendar html creation
javascript so that the popup code can be reused for ajax popups
in the future

fix bug that right-to-left language direction was not set based on
language settings selected by user in new tmpl files

fix bug that the stationery compose selection list appeared in the
bottom controlbar even if enable stationery was set to no

minor whitespace adjustments

NOTE TO TRANSLATORS:
Heavy modifications to the template files are required during this
transistion to separate logic from presentation. You will find that the
new .tmpl files are not translated. You may translate these files if you
like, however the next step in this transistion is to convert all
translations to the gettext .po format. The conversion to gettext .po
will force the loss of translations again. In short, if you translate
the new .tmpl files, thats great, but you will probably be asked to
translate again when we move to .po.

Please see the roadmap for an in-depth discussion of where the code is
going:

http://openwebmail.acatysmoof.com/dev/roadmap



git-svn-id: svn://openwebmail.acatysmoof.com/svn/trunk@391 9f51437f-89f6-0310-a5ff-ff9a3806d2ee