separate logic from presentation in openwebmail-saprefs.pl
NOTE TO TRANSLATORS:
Heavy modifications to the template files are required during this
transistion to separate logic from presentation. You will find that the
new .tmpl files are not translated. You may translate these files if you
like, however the next step in this transistion is to convert all
translations to the gettext .po format. The conversion to gettext .po
will force the loss of translations again. In short, if you translate
the new .tmpl files, thats great, but you will probably be asked to
translate again when we move to .po.

Please see the roadmap for an in-depth discussion of where the code is
going:

http://openwebmail.acatysmoof.com/dev/roadmap



git-svn-id: svn://openwebmail.acatysmoof.com/svn/trunk@426 9f51437f-89f6-0310-a5ff-ff9a3806d2ee