limit from name display to 40 characters in message listview
fix typo in openwebmail.conf.help

change the cardview popup display to take up the full popup area

change the wording of the image block bar to be remote images
instead of all images, since local images are still displayed

add new find-as-you-type autosuggest feature to the message
compose to, cc, and bcc fields. The users addressbook is
used to find suggestions to display that match the text
that is currently being typed. The typed text is matched
against a contacts full name, name, nickname, and email
address to find a suggestion match



git-svn-id: svn://openwebmail.acatysmoof.com/svn/trunk@494 9f51437f-89f6-0310-a5ff-ff9a3806d2ee